:recycle: 詳細ページの構造をロジックと表示部分に分割
